requirements.txt
Flask==2.2.5
pymongo==4.4.0
Werkzeug==2.2.3
Flask-Login==0.6.2
python-dotenv==1.0.0
